move claims to new package
