+++ packages/sdk/metadata-claims/src/domain.test.ts
@@ -3,8 +3,8 @@ import { newKitFromWeb3 } from '@celo/contractkit'
 import { testWithAnvilL1 } from '@celo/dev-utils/lib/anvil-test'
 import { ACCOUNT_ADDRESSES } from '@celo/dev-utils/lib/ganache-setup'
 import { NativeSigner, Signer, verifySignature } from '@celo/utils/lib/signatureUtils'
-import { IdentityMetadataWrapper } from '../metadata'
 import { DomainClaim, createDomainClaim, serializeClaim } from './claim'
+import { IdentityMetadataWrapper } from './metadata'
 import type { AccountMetadataSignerGetters } from './types'
 import { verifyDomainRecord } from './verify'
