+++ packages/sdk/metadata-claims/src/verify.ts
@@ -2,10 +2,10 @@ import { eqAddress } from '@celo/base/lib/address'
 import { isValidUrl } from '@celo/base/lib/io'
 import { resolveTxt } from 'node:dns' // TODO replace with non node specifc package
 import { promisify } from 'node:util' // TODO replace with non node specifc package
-import { IdentityMetadataWrapper } from '../metadata'
 import { AccountClaim } from './account'
 import { Claim, DOMAIN_TXT_HEADER, DomainClaim, serializeClaim } from './claim'
 import { verifyKeybaseClaim } from './keybase'
+import { IdentityMetadataWrapper } from './metadata'
 import { AccountMetadataSignerGetters, ClaimTypes } from './types'
 
 /**
